pyproject.toml
[tool.poetry]
name = "financial-analyst-qa-bot-rag-project"
version = "0.1.0"
description = "RAG & LLM End to End Project"
authors = ["Rajeev-Das <rajeebdas1998@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.10,<3.15"
streamlit = "^1.49.1"
pypdf2 = "^3.0.1"
sentence-transformers = "^5.1.0"
faiss-cpu = "^1.12.0"
openai = "^1.108.1"
huggingface-hub = "^0.35.0"
requests = "^2.32.5"
lxml = "^6.0.1"
pandas = "^2.3.2"
langchain = "^0.3.0"
langchain-openai = "^0.2.0"
langchain-community = "^0.3.0"
python-dotenv = "^1.0.0"
tiktoken = "^0.7.0"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.30.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
